rename celery_init.py
