+++ tests/test_celery.py
@@ -1,6 +1,6 @@
 from unittest.mock import patch
 from application.core.settings import settings
-from application.celery import make_celery
+from application.celery_init import make_celery
 
 
 @patch('application.celery.Celery')
